\name{data_seqgroups}
\alias{data_seqgroups}
\docType{data}
\title{data_seqgroups}
\description{A column vector of simulated data with 393 observations in 3 segments (which could, e.g., be groups or dyads).}
% \format{
  %
% }
\details{
A column vector of simulated data with 393 observations in 3 segments (which could, e.g., be groups or dyads). The beginning of each segment is indicated by a number greater than 999. The data set is provided as trial data for the seqgroups function.}
% \source{
% %%  ~~ reference to a publication or URL from which the data were obtained ~~
% }
% \references{
% }
\examples{
summary(data_seqgroups)
}
\keyword{dataset}
