\name{data_seqgroups_numeric}
\alias{data_seqgroups_numeric}
\docType{data}
\title{data_seqgroups_numeric}
\description{A column vector of simulated data with 393 observations in 3 segments 
	(which could, e.g., be groups or dyads).}
\details{
A column vector of numeric data with 393 observations in 3 segments (which could, 
e.g., be groups or dyads). The beginning of each segment is indicated by a number 
greater than 999. The data set is provided as trial data for the seqgroups function. 
It is a numeric version of the data in data_seqgroups_strings.}
\examples{
table(data_seqgroups_numeric)
}
\keyword{dataset}
