\name{data_seqgroups_strings}
\alias{data_seqgroups_strings}
\docType{data}
\title{data_seqgroups_strings}
\description{A column vector of simulated data with 393 observations in 3 segments (which could, e.g., be groups or dyads).}
\details{
A column vector of string data with 393 observations in 3 segments (which could, e.g., be groups or dyads). The beginning of each segment is indicated by the word "segment". The data set is provided as trial data for the seqgroups function. It is a string/character version of the data in data_seqgroups_numeric.}
\examples{
table(data_seqgroups_strings)
}
\keyword{dataset}
