\name{data_sequential}
\alias{data_sequential}
\docType{data}
\title{data_sequential}
\description{A column vector of trial data for sequential analyses.}
\details{
A column vector with 122 observations (codes). The data are provided as trial data for the sequential, bidirectional, twocells, paradom, and nonparadom functions.}
\examples{
table(data_sequential)
}
\keyword{dataset}
