\name{data_Wampold_1982}
\alias{data_Wampold_1982}
\docType{data}
\title{data_Wampold_1982}
\description{A vector of code sequences that mimic the frequency transition matrix and
	the statistical results reported in Wampold & Margolin (1982).}
\details{A column vector of 200 sequential codes. The data are provided as trial  
data for the paradom and nonparadom functions.}
\references{
 {Wampold, B. E., & Margolin, G. (1982). Nonparametric strategies to test 
	the independence of behavioral states in sequential data. \emph{Psychological 
	Bulletin, 92,} 755-765.}
}
\examples{
table(data_Wampold_1982)
}
\keyword{dataset}
