\name{data_Wampold_1984}
\alias{data_Wampold_1984}
\docType{data}
\title{data_Wampold_1984}
\description{A vector of code sequences that mimic the frequency transition matrix and
	the statistical results reported in Wampold (1984).}
\details{A column vector of 200 sequential codes. The data are provided as trial  
data for the paradom and nonparadom functions.}
\references{
 {Wampold, B. E. (1984). Tests of dominance in sequential categorical data. 
	\emph{Psychological Bulletin, 96,} 424-429.}
}
\examples{
table(data_Wampold_1984)
}
\keyword{dataset}
