\name{data_sequential}
\alias{data_sequential}
\docType{data}
\title{data_sequential}
\description{A column vector of hypothetical sequential data.}
% \format{
  %
% }
\details{
A column vector with 122 observations. The data set is provided as trial data for the sequential, bidirectional, twocells, paradom, and nonparadom functions.}
% \source{
% %%  ~~ reference to a publication or URL from which the data were obtained ~~
% }
% \references{
% }
\examples{
summary(data_sequential)
}
\keyword{dataset}
