\name{LagSequential-package}

\alias{LagSequential-package}

\docType{package}

\title{ Functions for analyzing lag-sequential categorical data }

\description{
This paper describes simple and flexible programs for conducting lag sequential event analyses using SAS and SPSS. The programs read a stream of codes and produce a variety of lag sequential statistics, including transitional frequencies, expected transitional frequencies, transitional probabilities, z values, adjusted residuals, Yule's Q values, likelihood ratio tests of stationarity across time and homogeneity across groups or segments, transformed kappas for unidirectional dependence, bidirectional dependence, parallel and nonparallel dominance, and significance levels based on both parametric and randomization tests.
}
\references{
 {O'Connor, B. P. (1999). Simple and flexible SAS and SPSS programs for analyzing lag-sequential categorical data. Behavior Research Methods, Instrumentation, and Computers, 31, 718-726.}
}

